Separate code to set formats/filters into different source files. This needs close_filters and free_filters to be globally available as global symbols __archive_read_close_filters and __archive_read_free_filters.
